Level saving + loading (basic serialization)
Firstly, moved all functionality from game into world and removed game as
game was a fairly useless class.

Added the cereal library to use with data serialization. Each actor/solid
we want to save can implement the IToRawData interface which returns a
polymorphic struct which should be able to recreate said actor/solid.
This struct implements the cereal serialization functionality to save
with. This implementation may change in the future.

Finished off the debug level menu to use interface with this new
functionality.